fix: Resolve batch 5 compile errors for TestFlight
- BioModulatorVisualization: Fix BioMetricsView init params
- BioParameterMapper: Add Float() conversions for Double→Float
- BioReactiveConfigurable: CoherenceMapping.linear→.standard
- BioReactiveShaderBridge: Shader.Argument.int→.float
- BiophysicalWellnessEngine: DispatchSourceTimer.invalidate()→.cancel()
- CameraManager: Add public to delegate methods, nonisolated for texture
- CoherenceTypes: Swift.max/Swift.min to avoid static member shadowing

+++ Sources/Echoelmusic/Shaders/BioReactiveShaderBridge.swift
@@ -94,7 +94,7 @@ extension ShaderLibrary {
         ShaderLibrary.default.coherenceParticles(
             .float(time),
             .float(coherence),
-            .int(Int32(particleCount))
+            .float(Float(particleCount))
         )
     }
 }
